Split rust library and binaries
