+++ Cargo.toml
@@ -1,8 +1,61 @@
 [workspace]
 resolver = "3"
 members = [
-    "source/symbolizer"
+    "source/rust/flamegraph",
+    "source/rust/memhawk_core",
+    "source/rust/symbolizer",
 ]
+
+[workspace.dependencies]
+memhawk_core = { path = "source/rust/memhawk_core" }
+
+# errors
+anyhow = "1.0.100"
+# rt
+tokio = { version = "1.47.1", features = ["rt", "rt-multi-thread"] }
+tokio-util = { version = "0.7.17", features = ["rt", "io"] }
+# logging
+env_logger = "0.11.8"
+log = "0.4.28"
+# command line argument parsing
+clap = { version = "4.5.51", features = ["derive"] }
+# clickhouse
+clickhouse = { version = "0.14.0", features = ["time"] }
+# time
+time = { version = "0.3.44", features = ["serde"] }
+# compression
+lz4 = "1.23.1"
+# postgres
+sqlx = { version = "0.8", features = ["runtime-tokio", "postgres", "macros"] }
+futures = "0.3.31"
+
+# saving state
+bitcode = "0.6.7" # N.B. format can change between major versions
+zstd = "0.13.3"
+sha2 = "0.10.9"
+
+# states
+serde_repr = "0.1.20"
+serde = { version = "1.0.228" }
+
+# protobufs
+bytes = { version = "1" }
+prost = "0.13.5"
+
+# enums
+strum = { version = "0.27.2", features = ["derive"] }
+derive_more = { version = "2.0.1", features = [
+    "std",
+    "from",
+    "into",
+    "add",
+    "add_assign",
+] }
+# symbol manager
+wholesym = "0.8.1"
+# hash table
+rustc-hash = "2.1.1"
+
 
 [profile.release]
 debug = true
