+++ source/rust/CMakeLists.txt
@@ -1,0 +1,8 @@
+
+find_program(CARGO cargo REQUIRED)
+
+add_subdirectory(flamegraph)
+# todo: is it necessary to add custom target?
+# add_subdirectory(memhawk_core)
+add_subdirectory(symbolizer)
+
