+++ source/rust/memhawk_core/src/proto/mod.rs
@@ -1,0 +1,1 @@
+pub mod schema;
