+++ source/rust/symbolizer/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "symbolizer"
+version = "0.1.0"
+edition = "2024"
+license = "Apache-2.0"
+
+[dependencies]
+memhawk_core.workspace = true
+anyhow.workspace = true
+bitcode.workspace = true
+rustc-hash.workspace = true
+strum.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
+env_logger.workspace = true
+log.workspace = true
+time.workspace = true
+clickhouse.workspace = true
+clap.workspace = true
+sqlx.workspace = true
+serde.workspace = true
+serde_repr.workspace = true
+sha2.workspace = true
+zstd.workspace = true
+prost.workspace = true
