+++ source/rust/symbolizer/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod clickhouse;
+pub mod postgres;
+pub mod processor;
+pub mod proto;
