+++ source/rust/symbolizer/src/proto/mod.rs
@@ -1,2 +1,1 @@
 pub mod reader;
-pub mod schema;
